remove expression/integration_serial_test
