--- expression/integration_serial_test/README.md
@@ -1,3 +1,0 @@
-Too many test case in this package.
-
-Please not add new test cases in this package.
