+++ session/nontransactionaltest/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "nontransactional_test.go",
     ],
     flaky = True,
-    shard_count = 21,
+    shard_count = 3,
     deps = [
         "//config",
         "//testkit",
